Idea of a new cache-based parsing
This implementation would allow parser combinators to be used (top-down
parsing) without the needles backtracking. When a new value is parsed it
would be stored in a data structure and if that parser fails, what it
had successfully parsed would remain in cache. Because of that, later
trials to parse the same value would just get it from cache, instead of
recalculating everithing. However, the current `std::any::TypId::of()`
restriction to only allow `'static` types prevents this approach from
working since every parsed structure would have to contain references
to other cached structures and those would live for shorter lifetimes.